Externalised relay (#308)
* refactor: update browserslist configuration and improve build process

* refactor: remove unused relay regex and add dynamic version fetching for React and Relay

* feat: externalise relay in production builds

+++ packages/gdu/commands/build/buildSPA.ts
@@ -42,7 +42,9 @@ export const buildSPA = async (guruConfig: GuruConfig) => {
 	);
 
 	// first clear build directory
-	await fs.rmdir(guruConfig.outputPath, { recursive: true });
+	if (await fs.stat(guruConfig.outputPath).catch(() => false)) {
+		await fs.rmdir(guruConfig.outputPath, { recursive: true });
+	}
 	await fs.mkdir(guruConfig.outputPath, { recursive: true });
 	const compiler = webpack(webpackConfigs);
 	await run(compiler);
